+++ nav2_system_tests/src/localization/test_localization_node.cpp
@@ -100,7 +100,7 @@ void TestAmclPose::initTestPose()
   testPose_.pose.pose.orientation.y = 0.0;
   testPose_.pose.pose.orientation.z = 0.0;
   testPose_.pose.pose.orientation.w = 1.0;
-  for (int i = 0; i < 35; i++) {
+  for (size_t i = 0; i < 35; i++) {
     testPose_.pose.covariance[i] = 0.0;
   }
   testPose_.pose.covariance[0] = 0.08;
